requirements.txt
apache-airflow==2.10.0
apache-airflow-providers-google
google-cloud-bigquery
uvicorn
fastapi